requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.4.6
APScheduler==3.10.1
lxml==4.9.3
